pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llmock"
dynamic = ["version"]
description = "Zero-cost local mock server for LLM API resilience testing"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
]

[project.scripts]
llmock = "llmock.cli:app"

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "twine>=5.1.1",
]

[tool.hatch.version]
path = "llmock/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["llmock"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]
